Refactor API controllers to move to backOffice namespace and update author handling
